requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2
decorator==4.4.2
Flask==1.1.2
idna==2.10
install==1.3.4
itsdangerous==1.1.0
jdatetime==3.6.2
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
protobuf==3.13.0
pycparser==2.20
PyMySQL==0.10.1
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
telegram==0.0.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
